@@ -1,9 +1,8 @@
 {
 	"require": {
-    	"php": ">=5.3.0",
+    	"php": ">=5.4.0",
 	    "ext-curl": "*",
 	    "ext-json": "*",
-	    "phpmailer/phpmailer": "~5.2",
 	    "smarty/smarty": "~3.1"
 	},
     "require-dev": {
