@@ -1,4 +1,6 @@
-<?php return array (
+<?php
+
+return array(
   'contact_email' => '',
   'contact_name' => '',
   'contact_adress' => '',
@@ -6,5 +8,5 @@
   'contact_postal_code' => '',
   'contact_region' => '',
   'contact_country' => '',
-  'contact_city' => ''
-) ;
+  'contact_city' => '',
+);
