@@ -1,4 +1,11 @@
-<?php return array (
-  'enable_composer' => true,
-  'composer_autoloader' => ''
-) ;
+<?php
+
+return array(
+  'enable_composer' 		=> true,
+  'enable_modules' 			=> true,
+  'enable_events' 			=> true,
+  'composer_autoloader' 	=> '',
+  'registry_caching'		=> false,
+  'registry_caching_method'	=> 'file',
+  'registry_caching_time'	=> 300,
+);
