@@ -1,9 +1,34 @@
-<?php return array (
-  'type' => '',
-  'host' => '',
-  'database' => '',
-  'username' => '',
-  'password' => '',
-  'prefix' => '',
-  'debug' => false,
-) ;
+<?php
+
+return array(
+	'active_group' 	=> 'default',
+	'query_builder' => true,
+	'databases' 	=> array(
+
+			'default' 		=> array(
+				'dsn'	=> '',
+				'hostname' => '',
+				'username' => '',
+				'password' => '',
+				'database' => '',
+				'dbdriver' => 'mysqli',
+				'dbprefix' => '',
+				'pconnect' => FALSE,
+				'db_debug' => FALSE,
+				'cache_on' => FALSE,
+				'cachedir' => '',
+				'char_set' => 'utf8',
+				'dbcollat' => 'utf8_general_ci',
+				'swap_pre' => '',
+				'encrypt' => FALSE,
+				'compress' => FALSE,
+				'stricton' => FALSE,
+				'failover' => array(),
+				'save_queries' => TRUE
+			),
+
+
+		),
+);
+
+
