@@ -1,4 +1,6 @@
-<?php return array (
+<?php
+
+return array(
   'debug' => false,
   'error_reporting' => true,
-) ;
+);
