@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -19,59 +19,64 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
- * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
- * @link        http://fuzeworks.techfuze.net
- * @since       Version 0.0.1
- * @version     Version 0.0.1
+ * @author    TechFuze
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
+ * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
+ * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
+ * @link  http://fuzeworks.techfuze.net
+ * @since Version 0.0.1
+ *
+ * @version Version 0.0.1
  */
 
 namespace FuzeWorks\Event;
-use \FuzeWorks\Event;
+
+use FuzeWorks\Event;
 
 /**
  * Event that gets loaded when a view is loaded.
  *
  * Use this to cancel the loading of a view, or change the file or engine of a view
  *
- * @package     net.techfuze.fuzeworks.core.event
- * @author      Abel Hoogeveen <abel@techfuze.net>
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * @author    Abel Hoogeveen <abel@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  */
-class LayoutLoadViewEvent extends Event {
-
-	/**
-	 * The directory of the view to be loaded
-	 * @var string
-	 */
+class LayoutLoadViewEvent extends Event
+{
+    /**
+     * The directory of the view to be loaded.
+     *
+     * @var string
+     */
     public $directory;
 
     /**
-     * The file of the view to be loaded
+     * The file of the view to be loaded.
+     *
      * @var string
      */
     public $file;
 
     /**
-     * The engine the file will be loaded with
+     * The engine the file will be loaded with.
+     *
      * @var object
      */
     public $engine;
 
     /**
-     * The assigned variables to the template
+     * The assigned variables to the template.
+     *
      * @var array
      */
     public $assigned_variables;
 
-    public function init($file, $directory, $engine, $assigned_variables){
+    public function init($file, $directory, $engine, $assigned_variables)
+    {
         $this->file = $file;
         $this->directory = $directory;
         $this->engine = $engine;
         $this->assigned_variables = $assigned_variables;
     }
 }
-
-?>
