@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -19,19 +19,21 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
- * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
- * @link        http://fuzeworks.techfuze.net
- * @since       Version 0.0.1
- * @version     Version 0.0.1
+ * @author    TechFuze
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
+ * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
+ * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
+ * @link  http://fuzeworks.techfuze.net
+ * @since Version 0.0.1
+ *
+ * @version Version 0.0.1
  */
 
 namespace FuzeWorks;
 
 /**
- * Class EventPriority
+ * Class EventPriority.
  *
  * The EventPriority is an "enum" which gives priorities an integer value, the higher the integer value, the lower the
  * priority. The available priorities are, from highest to lowest:
@@ -42,42 +44,41 @@ namespace FuzeWorks;
  * EventPriority::NORMAL
  * EventPriority::LOW
  * EventPriority::LOWEST
- * @package     net.techfuze.fuzeworks.core
- * @author      Abel Hoogeveen <abel@techfuze.net>
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ *
+ * @author    Abel Hoogeveen <abel@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  */
-
 abstract class EventPriority
 {
-
-    const LOWEST    = 5;
-    const LOW       = 4;
-    const NORMAL    = 3;
-    const HIGH      = 2;
-    const HIGHEST   = 1;
-    const MONITOR   = 0;
+    const LOWEST = 5;
+    const LOW = 4;
+    const NORMAL = 3;
+    const HIGH = 2;
+    const HIGHEST = 1;
+    const MONITOR = 0;
 
     /**
-     * Returns the string of the priority based on the integer
+     * Returns the string of the priority based on the integer.
+     *
      * @param $intPriorty
+     *
      * @return bool|string A bool when the integer isn't a priority. If the integer is a priority, the name is returned
      */
-    static function getPriority($intPriorty){
-
-        switch($intPriorty){
-
+    public static function getPriority($intPriorty)
+    {
+        switch ($intPriorty) {
             case 5:
-                return "EventPriority::LOWEST";
+                return 'EventPriority::LOWEST';
             case 4:
-                return "EventPriority::LOW";
+                return 'EventPriority::LOW';
             case 3:
-                return "EventPriority::NORMAL";
+                return 'EventPriority::NORMAL';
             case 2:
-                return "EventPriority::HIGH";
+                return 'EventPriority::HIGH';
             case 1:
-                return "EventPriority::HIGHEST";
+                return 'EventPriority::HIGHEST';
             case 0:
-                return "EventPriority::MONITOR";
+                return 'EventPriority::MONITOR';
             default:
                 return false;
         }
@@ -85,21 +86,23 @@ abstract class EventPriority
 
     /**
      * Returns the highest priority
-     * This function is needed for the firing of events in the right order,
+     * This function is needed for the firing of events in the right order,.
+     *
      * @return int
      */
-    static function getHighestPriority(){
-
-        return EventPriority::MONITOR;
+    public static function getHighestPriority()
+    {
+        return self::MONITOR;
     }
 
     /**
      * Returns the lowest priority
-     * This function is needed for the firing of events in the right order,
+     * This function is needed for the firing of events in the right order,.
+     *
      * @return int
      */
-    static function getLowestPriority(){
-
-        return EventPriority::LOWEST;
+    public static function getLowestPriority()
+    {
+        return self::LOWEST;
     }
 }
