@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -19,85 +19,94 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
- * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
- * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
- * @link        http://fuzeworks.techfuze.net
- * @since       Version 0.0.1
- * @version     Version 0.0.1
+ * @author    TechFuze
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
+ * @copyright Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
+ * @license   http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
+ * @link  http://fuzeworks.techfuze.net
+ * @since Version 0.0.1
+ *
+ * @version Version 0.0.1
  */
 
 namespace FuzeWorks;
 
 /**
  * Interface for a Module that gives abstract model types
- * A model server must contain the methods from this interface in order to correctly serve models
- * @package     net.techfuze.fuzeworks.core
- * @author      Abel Hoogeveen <abel@techfuze.net>
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * A model server must contain the methods from this interface in order to correctly serve models.
+ *
+ * @author    Abel Hoogeveen <abel@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  */
-interface ModelServer {
-	public function giveModel($type);
+interface ModelServer
+{
+    public function giveModel($type);
 }
 
 /**
- * Abstract class Model
+ * Abstract class Model.
  *
  * Abstract for a model data representation, loads the correct parent type
- * @package     net.techfuze.fuzeworks.core
- * @author      Abel Hoogeveen <abel@techfuze.net>
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ *
+ * @author    Abel Hoogeveen <abel@techfuze.net>
+ * @copyright Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  */
-abstract class Model {
+abstract class Model
+{
+    /**
+     * The parent class holder object
+     * Requests get redirected to this class.
+     *
+     * @var Parent Object
+     */
+    private $parentClass;
 
-	/**
-	 * The parent class holder object
-	 * Requests get redirected to this class
-	 * @access private
-	 * @var Parent Object
-	 */
-	private $parentClass;
+    /**
+     * Set the type of this model. Eg, use techfuze/databasemodel and Databasemodel to get a SQL connected model.
+     *
+     * @param string Module_name, the name of the module where the model can be found
+     * @param string Model_type, model type to return
+     */
+    protected function setType($module_name, $model_type)
+    {
+        $mod = Modules::get($module_name);
+        $this->parentClass = $mod->giveModel($model_type);
+    }
 
-	/**
-	 * Set the type of this model. Eg, use techfuze/databasemodel and Databasemodel to get a SQL connected model
-	 * @access protected
-	 * @param String Module_name, the name of the module where the model can be found
-	 * @param String Model_type, model type to return
-	 */
-	protected function setType($module_name, $model_type) {
-		$mod = Modules::get($module_name);
-		$this->parentClass = $mod->giveModel($model_type);
-	}
+    /**
+     * Retrieves a value from the model class.
+     *
+     * @param Any key
+     *
+     * @return Any value from the model class
+     */
+    public function __get($name)
+    {
+        return $this->parentClass->$name;
+    }
 
-	/**
-	 * Retrieves a value from the model class
-	 * @access public
-	 * @param Any key
-	 * @return Any value from the model class
-	 */
-	public function __get($name) {
-		return $this->parentClass->$name;
-	}
+    /**
+     * Sets a value in the model class.
+     *
+     * @param Any key
+     * @param Any value
+     */
+    public function __set($name, $value)
+    {
+        $this->parentClass->$name = $value;
+    }
 
-	/**
-	 * Sets a value in the model class
-	 * @access public
-	 * @param Any key
-	 * @param Any value
-	 */
-	public function __set($name, $value) {
-		$this->parentClass->$name = $value;
-	}
-
-	/**
-	 * Calls a function in the model class
-	 * @access public
-	 * @param String function_name
-	 * @param Array values
-	 * @return Function return
-	 */
-	public function __call($name, $params) {
-		return call_user_func_array(array($this->parentClass, $name), $params);
-	}
+    /**
+     * Calls a function in the model class.
+     *
+     * @param string function_name
+     * @param array values
+     *
+     * @return Function return
+     */
+    public function __call($name, $params)
+    {
+        return call_user_func_array(array($this->parentClass, $name), $params);
+    }
 }
