@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -20,24 +20,23 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * @copyright   Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
  * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
  * @link        http://fuzeworks.techfuze.net
  * @since       Version 0.0.1
+ *
  * @version     Version 0.0.1
  */
-
-use \FuzeWorks\Core;
-use \FuzeWorks\Router;
+use FuzeWorks\Core;
+use FuzeWorks\Router;
 
 // Include framework
-require_once( dirname(__FILE__) . "/Core/System/class.core.php");
+require_once dirname(__FILE__).'/Core/System/class.core.php';
 
 // Load it
 Core::init();
 
-Router::setPath( (isset($_GET['path']) ? $_GET['path'] : null)   );
+Router::setPath((isset($_GET['path']) ? $_GET['path'] : null));
 Router::route();
-
-?>
