@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -20,31 +20,32 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * @copyright   Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
  * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
  * @link        http://fuzeworks.techfuze.net
  * @since       Version 0.0.1
+ *
  * @version     Version 0.0.1
  */
-
 use FuzeWorks\Events;
 use FuzeWorks\Layout;
 
 /**
- * Class CoreTestAbstract
+ * Class CoreTestAbstract.
  *
  * Provides the event tests with some basic functionality
  */
 abstract class CoreTestAbstract extends PHPUnit_Framework_TestCase
 {
     /**
-     * Remove all listeners before the next test starts
+     * Remove all listeners before the next test starts.
      *
      * Reset the layout manager
      */
-    public function tearDown(){
-
+    public function tearDown()
+    {
         Events::$listeners = array();
         Layout::reset();
     }
