@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -20,11 +20,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * @copyright   Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
  * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
  * @link        http://fuzeworks.techfuze.net
  * @since       Version 0.0.1
+ *
  * @version     Version 0.0.1
  */
 // Load the abstract
@@ -32,8 +34,8 @@
 use \FuzeWorks\Config;
 use \FuzeWorks\Core;
 
-require_once "abstract.coreTestAbstract.php";
-require_once( "Core/System/class.core.php");
+require_once 'abstract.coreTestAbstract.php';
+require_once 'Core/System/class.core.php';
 
 ob_start();
 Core::init();
@@ -43,5 +45,3 @@ $cfg = Config::get('error');
 $cfg->debug = false;
 $cfg->error_reporting = false;
 $cfg->commit();
-
-?>
