@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -20,28 +20,29 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * @copyright   Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
  * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
  * @link        http://fuzeworks.techfuze.net
  * @since       Version 0.0.1
+ *
  * @version     Version 0.0.1
  */
-
 use \FuzeWorks\Core;
 use \FuzeWorks\Events;
 use \FuzeWorks\EventPriority;
 
 /**
- * Class CoreStartEventTest
+ * Class CoreStartEventTest.
  */
-class CoreStartEventTest extends CoreTestAbstract
+class coreStartEventTest extends CoreTestAbstract
 {
     /**
-     * Check if the event is fired when it should be
+     * Check if the event is fired when it should be.
      */
-    public function testCoreStartEvent(){
-
+    public function testCoreStartEvent()
+    {
         $mock = $this->getMock('MockEvent', array('mockMethod'));
         $mock->expects($this->once())->method('mockMethod');
 
