@@ -1,6 +1,6 @@
 <?php
 /**
- * FuzeWorks
+ * FuzeWorks.
  *
  * The FuzeWorks MVC PHP FrameWork
  *
@@ -20,28 +20,29 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  * @author      TechFuze
- * @copyright   Copyright (c) 2013 - 2015, Techfuze. (http://techfuze.net)
+ * @copyright   Copyright (c) 2013 - 2016, Techfuze. (http://techfuze.net)
  * @copyright   Copyright (c) 1996 - 2015, Free Software Foundation, Inc. (http://www.fsf.org/)
  * @license     http://opensource.org/licenses/GPL-3.0 GPLv3 License
+ *
  * @link        http://fuzeworks.techfuze.net
  * @since       Version 0.0.1
+ *
  * @version     Version 0.0.1
  */
-
 use \FuzeWorks\Events;
 use \FuzeWorks\Router;
 use \FuzeWorks\EventPriority;
 
 /**
- * Class RouterRouteEventTest
+ * Class RouterRouteEventTest.
  */
-class RouterRouteEventTest extends CoreTestAbstract{
-
+class routerRouteEventTest extends CoreTestAbstract
+{
     /**
-     * Check if the event is fired when it should be
+     * Check if the event is fired when it should be.
      */
-    public function test_basic(){
-
+    public function test_basic()
+    {
         $mock = $this->getMock('MockEvent', array('mockMethod'));
         $mock->expects($this->once())->method('mockMethod');
 
@@ -51,10 +52,10 @@ class RouterRouteEventTest extends CoreTestAbstract{
     }
 
     /**
-     * Cancel events
+     * Cancel events.
      */
-    public function test_cancel(){
-
+    public function test_cancel()
+    {
         Router::setPath('x/y/z');
 
         Events::addListener(array($this, 'listener_cancel'), 'routerRouteEvent', EventPriority::NORMAL);
@@ -66,8 +67,8 @@ class RouterRouteEventTest extends CoreTestAbstract{
     }
 
     // Cancel all calls
-    public function listener_cancel($event){
-
+    public function listener_cancel($event)
+    {
         $event->setCancelled(true);
     }
 }
